# src/assets/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides a REST API for the Pool Project",
        "version": "1.0.0",
        "title": "Pool REST API",
        "contact": {
            "name": "Pool project team",
            "url": "https://github.com/no-domestic-violence"
        }
    },
    "host": "pool-api-mobile.herokuapp.com/api",
    "basePath": "/",
...
    "paths": {
        "/articles": {
            "post": {
                "tags": [
                    "article-controller"
                ],
                "summary": "Create article",
                "operationId": "createArticleUsingPOST",
                "consumes": "application/json",
                "produces":  "application/json",
                "parameters": [
                    {
                        "in": "body",
                        "name": "request",
                        "description": "request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Article"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/articlesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            },
            "get": {
                "tags": [
                    "article-controller"
                ],
                "summary": "Get all articles",
                "operationId": "getAllArticlesUsingGET",
                "produces": 
                    "application/json"
                ,
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Article"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false

            }
        },
        "/articles/{articleId}": {
            "get": {
                "tags": [
                    "article-controller"
                ],
                "summary": "Get article by Id",
                "operationId": "getArticleByIdUsingGET",
                "produces": 
                    "application/json",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Id of article that needs to be fetched",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema":  {
                            "$ref": "#/definitions/Article"
                        }
                    },
                    "400": {
                        "description": "Invalid article Id"
                    },
                    "404": {
                        "description": "Article not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "article-controller"
                ],
                "summary": "Delete article by Id",
                "operationId": "deleteArticleByIdUsingDELETE",
                "produces": 
                    "application/json",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "Article id to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Article successfully deleted"
                    },
                    "400": {
                        "description": "Invalid article Id"
                    },
                    "404": {
                        "description": "Article not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Article": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "url_to_image": {
                    "type": "string"
                },
                "violence_type": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "Sexual",
                            "Financial",
                            "Emotional"
                        ]
                    }
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                }
            },
            "title": "Article"
        },
...
        "articlesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "$ref": "#/definitions/Article"
                },
                "success": {
                    "type": "boolean"
                }
            },
            "title": "articlesResponse"
        }
    },
    "schemes": [
        "https"
    ]
}